REFACTOR: Js Linting
