+++ assets/src/admin/media-sharing/components/syncIcon.js
