+++ assets/src/admin/media-sharing/components/versionIcon.js
