--- assets/src/js/admin.js
@@ -1,3 +1,0 @@
-/**
- * Internal dependencies
- */
