--- assets/src/js/editor.js
@@ -1,3 +1,0 @@
-/**
- * This will be loaded in the editor side only.
- */
