--- assets/src/js/main.js
@@ -1,3 +1,0 @@
-/**
- * Main JavaScript file for the OneMedia plugin.
- */
