+++ package.json
@@ -45,6 +45,7 @@
 		"eslint-plugin-eslint-comments": "^3.2.0",
 		"eslint-plugin-import": "^2.32.0",
 		"eslint-plugin-jest": "^27.9.0",
+		"eslint-plugin-react-hooks": "^7.0.1",
 		"globals": "^16.5.0",
 		"npm-run-all": "^4.1.5",
 		"rimraf": "^6.1.2",
